services/api/requirements.txt
fastapi==0.116.1
httpx==0.28.1
openai==1.107.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
uvicorn[standard]==0.35.0

# tests
pytest==8.4.2
